test/requirements.txt
ExifRead==2.3.2
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
progressbar2==3.53.1
python-utils==2.4.0
PyWavelets==1.1.1
scipy==1.6.0
six==1.15.0
matplotlib==3.3.3
